# design-handoff/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "메뉴 관리 API 명세서",
    "description": "Spring Data JPA 수업용 프로젝트(chap06)의 API 명세서다. 메뉴와 카테고리 조회, 메뉴 등록·수정·삭제를 제공한다.",
    "contact": {
      "name": "Ohgiraffers 강사팀",
      "email": "owl.ohgiraffers@gmail.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "메뉴",
      "description": "메뉴 조회와 등록·수정·삭제를 제공한다."
    },
    {
      "name": "카테고리",
      "description": "메뉴가 속한 카테고리를 조회한다."
    }
  ],
  "paths": {
    "/api/menus/{menuCode}": {
      "get": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 상세 조회",
        "description": "메뉴 코드로 메뉴 한 건을 조회한다. 응답의 result.menu 에 MenuDTO 가 담긴다.",
        "operationId": "findMenuByCode",
        "parameters": [
          {
            "name": "menuCode",
            "in": "path",
            "description": "조회할 메뉴의 코드(PK)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "해당 코드의 메뉴가 없음"
          }
        }
      },
      "put": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 수정",
        "description": "메뉴 코드에 해당하는 메뉴를 수정한다. 요청 본문은 등록과 같은 모양이고, 바꾸지 않을 값도 함께 보낸다. 응답의 result.menu 에 수정된 MenuDTO 가 담긴다.",
        "operationId": "updateMenu",
        "parameters": [
          {
            "name": "menuCode",
            "in": "path",
            "description": "수정할 메뉴의 코드(PK)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "해당 코드의 메뉴가 없음"
          }
        }
      },
      "delete": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 삭제",
        "description": "메뉴 코드에 해당하는 메뉴를 삭제한다. **응답 상태는 200 이다.** 템플릿 안의 httpStatus 에는 204 가 적혀 있지만, 204 는 본문을 가질 수 없어 실제 응답은 200 으로 내려간다. 응답의 result.deletedMenuCode 에 삭제된 코드가 담긴다.",
        "operationId": "deleteMenu",
        "parameters": [
          {
            "name": "menuCode",
            "in": "path",
            "description": "삭제할 메뉴의 코드(PK)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "해당 코드의 메뉴가 없음"
          }
        }
      }
    },
    "/api/menus": {
      "get": {
        "tags": [
          "메뉴"
        ],
        "summary": "전체 메뉴 조회",
        "description": "등록된 모든 메뉴를 페이징 없이 조회한다. 응답의 result.menus 에 MenuDTO 배열이 담긴다.",
        "operationId": "findAllMenus",
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 등록",
        "description": "새 메뉴를 등록한다. 요청 본문에는 menuName, menuPrice, categoryCode, orderableStatus 를 담는다. menuCode 는 서버가 채우므로 보내지 않는다. orderableStatus 는 'Y' 또는 'N' 한 글자다. 응답의 result.menu 에 등록된 MenuDTO 가 담긴다.",
        "operationId": "saveMenu",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "등록 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "400": {
            "description": "요청 본문이 올바르지 않음"
          }
        }
      }
    },
    "/api/menus/search": {
      "get": {
        "tags": [
          "메뉴"
        ],
        "summary": "가격으로 메뉴 검색",
        "description": "지정한 가격을 **초과하는**(미만이 아니다) 메뉴를 조회한다. 응답의 result.menus 에 MenuDTO 배열이, result.searchPrice 에 기준 가격이 담긴다.",
        "operationId": "findMenusByPrice",
        "parameters": [
          {
            "name": "menuPrice",
            "in": "query",
            "description": "기준 가격. 이 값을 초과하는 메뉴만 조회된다",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/menus/pages": {
      "get": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 목록 조회 (페이징)",
        "description": "페이지 단위로 메뉴를 조회한다. 요청 파라미터는 page 와 size 이고, **page 는 1부터 센다.** 응답의 result 에 content(MenuDTO 배열), totalElements, totalPages, size, number(현재 페이지, 1부터), first, last 가 담긴다.",
        "operationId": "findMenuPage",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/menus/pages/sort": {
      "get": {
        "tags": [
          "메뉴"
        ],
        "summary": "메뉴 목록 조회 (페이징 + 정렬)",
        "description": "정렬 기준을 지정해 페이지 단위로 메뉴를 조회한다. page 는 1부터 세고, 기본 페이지 크기는 5다. 응답의 result 는 위 페이징 조회와 같고, sort 와 direction 이 더 담긴다.",
        "operationId": "findMenuPageWithSort",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "정렬 기준 필드 (menuCode, menuName, menuPrice 등)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "menuPrice"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "정렬 방향. asc 또는 desc",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "카테고리"
        ],
        "summary": "전체 카테고리 조회",
        "description": "모든 카테고리를 조회한다. 메뉴 등록·수정 폼의 선택 목록과 목록 화면의 필터에 쓴다. 응답의 result.categories 에 CategoryDTO 배열이 담긴다. 카테고리는 상위-하위 두 단계다. refCategoryCode 와 refCategoryName 이 상위 카테고리를 가리키고, 최상위 카테고리(식사·음료·디저트)는 둘 다 null 이다. 메뉴가 실제로 속하는 것은 하위 카테고리다.",
        "operationId": "findAllCategories",
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{categoryCode}": {
      "get": {
        "tags": [
          "카테고리"
        ],
        "summary": "카테고리 상세 조회",
        "description": "카테고리 코드로 카테고리 한 건을 조회한다. 응답의 result.category 에 CategoryDTO 가 담긴다.",
        "operationId": "findCategoryByCode",
        "parameters": [
          {
            "name": "categoryCode",
            "in": "path",
            "description": "조회할 카테고리의 코드(PK)",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "해당 코드의 카테고리가 없음"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MenuDTO": {
        "type": "object",
        "properties": {
          "menuCode": {
            "type": "integer",
            "format": "int32"
          },
          "menuName": {
            "type": "string"
          },
          "menuPrice": {
            "type": "integer",
            "format": "int32"
          },
          "categoryCode": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string"
          },
          "orderableStatus": {
            "type": "string"
          }
        }
      },
      "ResponseMessage": {
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}